Factor out txn.rs from mod.rs in persistence lmdb
